Removed the old gait selection interface, added another file for the separate pop up class and removed the findchild functions because the children are attributes of the class and can be called as such.
